git-svn-id: http://citadel.physiology.northwestern.edu/svn/s1_analysis@888 984b1e3c-1c57-0410-ae62-c18f90d3b645

+++ proc/robert/buildModel_fp.m
@@ -163,7 +163,7 @@ if exist('words','var') && ~isempty(words)
 else
     q=ones(1,length(analog_times));   %Temp kludge b/c find_active_regions gives too long of a vector back
 end
-q = interp1(analog_times, q, t);
+q = interp1(analog_times, double(q), t);
 
 disp('2nd part:assign t,y,q')
 toc
